requirements.txt
python-telegram-bot==20.7
openai==0.28.0
flask
yookassa
textblob
